DirectX 12テクスチャ管理のリファクタリングと機能拡張
Dx12ResourceFactoryにテクスチャ・深度バッファ生成関数を追加し、リソース生成処理を集約。TextureManagerは重複読み込み防止のためのパス→IDマップを導入し、ID管理方法をvectorベースに変更。SRV_UAVManagerはTexMetadataを利用したSRV生成に対応。SwapChainManagerの深度バッファ生成もFactory経由に統一。不要コード削除やインクルード整理など全体的な可読性・保守性も向上。

+++ DirectXGame/src/engine/DrawSystem/DrawSystem.cpp
@@ -91,40 +91,21 @@ void DrawSystem::DrawRenderObject()
 
 		// 4) CBV・SRVセット
 		const auto& cpuStrage = renderObject->GetCpuStorage();
-		const auto& rootParams = renderObject->GetRootParams();
-
-		//uint32_t 
+		const auto& rootParams = renderObject->GetRootParams(); 
 		for (const auto& [key, param] : rootParams)
 		{
 			if (param.paramType == ParamType::CBV)
 			{
 				const auto alloc = cb.Allocate(param.sizeBytes);
 				std::memcpy(alloc.cpu, cpuStrage.data() + param.offsetBytes, param.sizeBytes);
-				cmdList->SetGraphicsRootConstantBufferView(param.vectorIndex, alloc.gpu);
+				cmdList->SetGraphicsRootConstantBufferView(UINT(param.vectorIndex), alloc.gpu);
 			}
 			else if (param.paramType == ParamType::SRV)
 			{
 				assert(param.srvAllocIndex != UINT32_MAX);
-				cmdList->SetGraphicsRootDescriptorTable(param.vectorIndex, srvManager->GetGPUHandleAt(param.srvAllocIndex));
+				cmdList->SetGraphicsRootDescriptorTable(UINT(param.vectorIndex), srvManager->GetGPUHandleAt(param.srvAllocIndex));
 			}
 		}
-
-		//for (size_t i = 0; i < rootParams.size(); ++i)
-		//{
-		//	const auto& param = rootParams[i];
-		//
-		//	if (param.paramType == ParamType::CBV)
-		//	{
-		//		const auto alloc = cb.Allocate(param.sizeBytes);
-		//		std::memcpy(alloc.cpu, cpuStrage.data() + param.offsetBytes, param.sizeBytes);
-		//		cmdList->SetGraphicsRootConstantBufferView(static_cast<UINT>(i), alloc.gpu);
-		//	}
-		//	else if (param.paramType == ParamType::SRV)
-		//	{
-		//		assert(param.srvAllocIndex != UINT32_MAX);
-		//		cmdList->SetGraphicsRootDescriptorTable(static_cast<UINT>(i), srvManager->GetGPUHandleAt(param.srvAllocIndex));
-		//	}
-		//}
 
 		// モデルの検索
 		const ModelData* obj = resourceManager_->GetModelManager()->GetModelData(renderObject->modelID);
